Avatars can now be 32 bit sizes. (see https://xmpp.org/extensions/xep-0084.html) size can be xs:unsignedInt

+++ parsers/src/avatar.rs
@@ -24,7 +24,7 @@ generate_element!(
     Info, "info", AVATAR_METADATA,
     attributes: [
         /// The size of the image data in bytes.
-        bytes: Required<u16> = "bytes",
+        bytes: Required<u32> = "bytes",
 
         /// The width of the image in pixels.
         width: Option<u16> = "width",
